Extract toolbar refresh helper for use in StringTable and Plots
