+++ core/ui/src/main/java/org/phoebus/ui/javafx/StringTable.java
@@ -14,7 +14,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.concurrent.ForkJoinPool;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
@@ -483,18 +482,8 @@ public class StringTable extends BorderPane
         // Force layout to reclaim space used by hidden toolbar,
         // or make room for the visible toolbar
         layoutChildren();
-        // XX Hack: Toolbar is garbled, all icons in pile at left end,
-        // when shown the first time, i.e. it was hidden when the plot
-        // was first shown.
-        // Manual fix is to hide and show again.
-        // Workaround is to force another layout a little later
-        if (show)
-            ForkJoinPool.commonPool().submit(() ->
-            {
-                Thread.sleep(1000);
-                Platform.runLater(() -> layoutChildren() );
-                return null;
-            });
+
+        ToolbarHelper.refreshHack(toolbar);
     }
 
     /** @param color Background color */
