+++ core/ui/src/main/java/org/phoebus/ui/javafx/ToolbarHelper.java
@@ -7,7 +7,12 @@
  *******************************************************************************/
 package org.phoebus.ui.javafx;
 
+import java.util.concurrent.ForkJoinPool;
+
+import javafx.application.Platform;
 import javafx.scene.Node;
+import javafx.scene.control.ButtonBase;
+import javafx.scene.control.ToolBar;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.Region;
@@ -43,5 +48,41 @@ public class ToolbarHelper
         sep.setMinWidth(Region.USE_PREF_SIZE);
         sep.setMaxWidth(Region.USE_PREF_SIZE);
         return sep;
+    }
+
+    /** Hack ill-sized toolbar buttons
+     *
+     *  <p>When toolbar is originally hidden and then later
+     *  shown, it tends to be garbled, all icons in pile at left end,
+     *  Manual fix is to hide and show again.
+     *  Workaround is to force another layout a little later.
+     */
+    public static void refreshHack(final ToolBar toolbar)
+    {
+        if (toolbar.getParent() == null)
+            return;
+        for (Node node : toolbar.getItems())
+        {
+            if (! (node instanceof ButtonBase))
+                continue;
+            final ButtonBase button = (ButtonBase) node;
+            final Node icon = button.getGraphic();
+            if (icon == null)
+                continue;
+            // Re-set the icon to force new layout of button
+            button.setGraphic(null);
+            button.setGraphic(icon);
+            if (button.getWidth() == 0  ||  button.getHeight() == 0)
+            {   // If button has no size, yet, try again later
+                ForkJoinPool.commonPool().submit(() ->
+                {
+                    Thread.sleep(500);
+                    Platform.runLater(() -> refreshHack(toolbar));
+                    return null;
+                });
+                return;
+            }
+        }
+        Platform.runLater(() -> toolbar.layout());
     }
 }
